feat: implement OAuth in Gitea for drone
The script checks and compares if there is a secret on the cluster, and whether is matches with the
oauth in Gitea. If it maches, it checks whether authorization for that oauth key is granted, if not
grant it as well. That way human interaction is minimized.

redkubes/unassigned-issues#170, redkubes/otomi-core#364, #26